+++ pkg/parca/parca_test.go
@@ -139,7 +139,12 @@ func Benchmark_WriteRaw(b *testing.B) {
 	<-done
 }
 
-func replayDebugLog(ctx context.Context, t require.TestingT) (querypb.QueryServiceServer, *frostdb.Table, *semgroup.Group, func()) {
+type Testing interface {
+	require.TestingT
+	Helper()
+}
+
+func replayDebugLog(ctx context.Context, t Testing) (querypb.QueryServiceServer, *frostdb.Table, *semgroup.Group, func()) {
 	dir := "../../tmp/"
 	files, err := ioutil.ReadDir(dir)
 	require.NoError(t, err)
@@ -208,7 +213,8 @@ func replayDebugLog(ctx context.Context, t require.TestingT) (querypb.QueryServi
 		logger,
 	)
 	require.NoError(t, err)
-	m := metastore.NewBadgerMetastore(
+	m := metastore.NewTestMetastore(
+		t,
 		logger,
 		reg,
 		tracer,
@@ -321,6 +327,8 @@ func MustReadAllGzip(t require.TestingT, filename string) []byte {
 }
 
 func TestConsistency(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 	logger := log.NewNopLogger()
 	reg := prometheus.NewRegistry()
@@ -340,7 +348,8 @@ func TestConsistency(t *testing.T) {
 		logger,
 	)
 	require.NoError(t, err)
-	m := metastore.NewBadgerMetastore(
+	m := metastore.NewTestMetastore(
+		t,
 		logger,
 		reg,
 		tracer,