+++ pkg/query/flamegraph_flat_test.go
@@ -35,10 +35,13 @@ import (
 )
 
 func TestGenerateFlamegraphFlat(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 	var err error
 
-	l := metastore.NewBadgerMetastore(
+	l := metastore.NewTestMetastore(
+		t,
 		log.NewNopLogger(),
 		prometheus.NewRegistry(),
 		trace.NewNoopTracerProvider().Tracer(""),
@@ -202,10 +205,13 @@ func TestGenerateFlamegraphFlat(t *testing.T) {
 }
 
 func TestGenerateFlamegraphFromProfile(t *testing.T) {
+	t.Parallel()
+
 	tracer := trace.NewNoopTracerProvider().Tracer("")
 	reg := prometheus.NewRegistry()
 
-	l := metastore.NewBadgerMetastore(
+	l := metastore.NewTestMetastore(
+		t,
 		log.NewNopLogger(),
 		reg,
 		tracer,
@@ -237,12 +243,14 @@ func testGenerateFlamegraphFromProfile(t *testing.T, l metastorepb.MetastoreServ
 }
 
 func TestGenerateFlamegraphWithInlined(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 	logger := log.NewNopLogger()
 	reg := prometheus.NewRegistry()
 	tracer := trace.NewNoopTracerProvider().Tracer("")
 
-	store := metastore.NewBadgerMetastore(logger, reg, tracer)
+	store := metastore.NewTestMetastore(t, logger, reg, tracer)
 
 	functions := []*pprofprofile.Function{
 		{ID: 1, Name: "net.(*netFD).accept", SystemName: "net.(*netFD).accept", Filename: "net/fd_unix.go"},
@@ -292,7 +300,7 @@ func TestGenerateFlamegraphWithInlined(t *testing.T) {
 	require.Equal(t, &pb.Flamegraph{
 		Total:  1,
 		Height: 4,
-		Unit:   "",
+		Unit:   "bytes",
 		Root: &pb.FlamegraphRootNode{
 			Cumulative: 1,
 			Children: []*pb.FlamegraphNode{{
@@ -379,12 +387,14 @@ func TestGenerateFlamegraphWithInlined(t *testing.T) {
 }
 
 func TestGenerateFlamegraphWithInlinedExisting(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 	logger := log.NewNopLogger()
 	reg := prometheus.NewRegistry()
 	tracer := trace.NewNoopTracerProvider().Tracer("")
 
-	store := metastore.NewBadgerMetastore(logger, reg, tracer)
+	store := metastore.NewTestMetastore(t, logger, reg, tracer)
 	metastore := metastore.NewInProcessClient(store)
 
 	functions := []*pprofprofile.Function{