+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.12](https://github.com/parca-dev/parca/compare/ui-v0.13.11...ui-v0.13.12) (2022-06-27)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.13.11](https://github.com/parca-dev/parca/compare/ui-v0.13.10...ui-v0.13.11) (2022-06-23)
+
+**Note:** Version bump only for package @parca/components
+
 ## [0.13.10](https://github.com/parca-dev/parca/compare/ui-v0.13.9...ui-v0.13.10) (2022-06-22)
 
 **Note:** Version bump only for package @parca/components